flash-cards/python-flash-cards/requirements.txt
boto==2.40.0
click>=8.1.3
configparser==3.5.0
cryptography==42.0.4
enum34==1.1.6
Flask==2.3.2
future==0.18.3
httplib2==0.19.0
ipaddress==1.0.16
itsdangerous>=2.1.2
Jinja2==3.1.4
jsonschema==2.5.1
lockfile==0.12.2
MarkupSafe>=2.0
ndg-httpsclient==0.4.2
psutil==5.6.6
psycopg2-binary==2.9.7
pyasn1==0.1.9
pycurl==7.43.0
pyOpenSSL==17.5.0
pytz==2016.10
pyxdg==0.26
requests==2.32.0
scour==0.32
six==1.10.0
virtualenv==15.1.0
Werkzeug==3.0.3
